Dependency upgrade, update Roslynator.Analyzers and Testcontainers (#1074)
* chore(dependencies): update Roslynator.Analyzers and Testcontainers

- Roslynator.Analyzers has been updated from version 4.11.0 to 4.12.0
  across multiple projects.
- Testcontainers has been updated from version 3.7.0 to 3.8.0 in
  NetDaemon.Tests.Integration project.

* feat(tests): add logging to HomeAssistantLifetime

This commit adds a logger to the HomeAssistantLifetime class to
provide more detailed information during test execution. It also
removes the logger from the HomeAssistantContainer constructor
as it was not being used. Minor formatting changes were also made
to improve code readability.

BREAKING CHANGE: The HomeAssistantContainer constructor no longer
accepts a logger as a parameter.